+++ R/utils.R
@@ -2,3 +2,8 @@
 isValidAndUnreserved <- function(string) {
   make.names(string) == string
 }
+
+# Modify from https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r
+split_vector <- function(x, batch_size) {
+  split(x, ceiling(seq_along(x) / batch_size))
+}